requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
langchain==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.9
langchain-community==0.3.29
chromadb==0.5.3
pypdf==4.2.0
docx2txt==0.8